# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FlagLab API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suggested-plays:
    get:
      operationId: listSuggestedPlays
      tags: [suggested-plays]
      summary: List suggested plays catalog
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [5v5, 7v7]
        - name: coverage
          in: query
          schema:
            type: string
        - name: concept
          in: query
          schema:
            type: string
        - name: field_zone
          in: query
          schema:
            type: string
        - name: tag
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of suggested plays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SuggestedPlay"
  /suggested-plays/{id}/load:
    post:
      operationId: loadSuggestedPlay
      tags: [suggested-plays]
      summary: Load a suggested play into the user's library
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Play created from suggested play
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Play"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    PlayerToken:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          description: "QB, WR1, WR2, WR3, RB, C, DB, Rusher, Safety, LB, etc."
        label:
          type: string
        x:
          type: number
        y:
          type: number
        color:
          type: string
        team:
          type: string
          enum: [offense, defense]
      required:
        - id
        - role
        - label
        - x
        - y
        - team
    RoutePoint:
      type: object
      properties:
        x:
          type: number
        y:
          type: number
      required:
        - x
        - y
    RouteSegment:
      type: object
      properties:
        id:
          type: string
        playerId:
          type: string
        type:
          type: string
          enum: [straight, slant, out, in, post, corner, go, comeback, hitch, drag, wheel, motion, handoff, zone_drop, screen]
        points:
          type: array
          items:
            $ref: "#/components/schemas/RoutePoint"
        style:
          type: string
          enum: [solid, dashed, dotted]
        color:
          type: string
      required:
        - id
        - playerId
        - type
        - points
        - style
    Play:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        mode:
          type: string
          enum: [offense, defense]
        format:
          type: string
          enum: [5v5, 7v7]
        formationId:
          type: string
        playbookId:
          type: string
        players:
          type: array
          items:
            $ref: "#/components/schemas/PlayerToken"
        routes:
          type: array
          items:
            $ref: "#/components/schemas/RouteSegment"
        notes:
          type: string
        tags:
          type: array
          items:
            type: string
        suggestedUse:
          type: string
        coverageTargets:
          type: array
          items:
            type: string
        fieldZone:
          type: string
          enum: [red_zone, midfield, backed_up]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        primaryConcept:
          type: string
        isManBeater:
          type: boolean
        isZoneBeater:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - mode
        - format
        - players
        - routes
        - tags
        - coverageTargets
        - createdAt
        - updatedAt
...
    SuggestedPlay:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        coachingPurpose:
          type: string
        idealSituation:
          type: string
        format:
          type: string
          enum: [5v5, 7v7]
        mode:
          type: string
          enum: [offense, defense]
        tags:
          type: array
          items:
            type: string
        coverageAttacked:
          type: string
        primaryConcept:
          type: string
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
        fieldZone:
          type: string
          enum: [red_zone, midfield, backed_up, any]
        isManBeater:
          type: boolean
        isZoneBeater:
          type: boolean
        players:
          type: array
          items:
            $ref: "#/components/schemas/PlayerToken"
        routes:
          type: array
          items:
            $ref: "#/components/schemas/RouteSegment"
        formationId:
          type: string
      required:
        - id
        - name
        - coachingPurpose
        - idealSituation
        - format
        - mode
        - tags
        - players
        - routes
...