# lib/api-zod/src/generated/types/suggestedPlay.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlagLab API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerToken } from "./playerToken";
import type { RouteSegment } from "./routeSegment";
import type { SuggestedPlayDifficulty } from "./suggestedPlayDifficulty";
import type { SuggestedPlayFieldZone } from "./suggestedPlayFieldZone";
import type { SuggestedPlayFormat } from "./suggestedPlayFormat";
import type { SuggestedPlayMode } from "./suggestedPlayMode";

export interface SuggestedPlay {
  id: string;
  name: string;
  coachingPurpose: string;
  idealSituation: string;
  format: SuggestedPlayFormat;
  mode: SuggestedPlayMode;
  tags: string[];
  coverageAttacked?: string;
  primaryConcept?: string;
  difficulty?: SuggestedPlayDifficulty;
  fieldZone?: SuggestedPlayFieldZone;
  isManBeater?: boolean;
  isZoneBeater?: boolean;
  players: PlayerToken[];
  routes: RouteSegment[];
  formationId?: string;
}